requirements.txt
# Core dependencies
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
pyyaml>=6.0.0

# MQTT client (async)
aiomqtt>=2.0.0

# HTTP client (async)
aiohttp>=3.9.0

# Protobuf / Meshtastic parsing
meshtastic>=2.4.0
protobuf>=4.24.0

# Type hints
typing-extensions>=4.8.0

